Added some new exercises
Added exercises 3.7, 3.8, 3.9 and 4.1 and cleaned up code.

+++ MyProject/MyProject/Topic2.h
@@ -10,7 +10,7 @@ class Topic2 {
 public:
 	void selectExercise();
 private:
-	int exercise;
+	short exercise;
 
 	void betterLinEqSys();
 	void testTriangle();
